marketpro/requirement.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
pandas==2.1.1
plotly==5.16.0
matplotlib==3.8.0
seaborn==0.12.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0